service/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "Placement Controller"
version = "0.0.0"
description = "Baseline placement controller for decentralized control plane."
readme = "README.md"
authors = ["HIRO-MicroDataCenters <all-hiro@hiro-microdatacenters.nl>"]
license = "MIT"
repository = "https://github.com/HIRO-MicroDataCenters-BV/placement-controller"
homepage = "https://github.com/HIRO-MicroDataCenters-BV/placement-controller"
packages = [{include = "*", from="src"}]

[tool.poetry.dependencies]
python = "^3.11.4"
kubernetes = "^29.0.0"
jsonpath = "^0.82.2"
jsonpath-ng = "^1.6.1"
httpx = "^0.27.0"
pydantic-settings = "^2.2.1"
aiohttp = "3.9.3"
loguru = "0.7.2"
kubernetes_asyncio = "30.1.1"
sortedcontainers = "2.4.0"
prometheus-async = "22.2.0"
fastapi = "0.115.12"
uvicorn = "0.34.3"

[tool.poetry.group.dev.dependencies]
black = "^23.12"
flake8 = "^7.0"
isort = {extras = ["colors"], version = "^5.13"}
mypy = "^1.8"
mypy-extensions = "^1.0"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.12"
pre-commit = "^3.7.0"
types-python-dateutil = "^2.9.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4"
pytest-mock = "^3.12.0"
vcrpy = "6.0.1"

[tool.isort]
# https://github.com/timothycrosley/isort/
line_length = 120

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.black]
# https://github.com/psf/black
line-length = 120
color = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
# mypy_path = "/"
pretty = true
show_traceback = true
color_output = true
exclude = [
  '.git',
  '__pycache__',
  '.venv',
  'build',
  'dist'
]

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]
